+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
+## 1.2.0
+- Added AudioInstance in JoyDO.js and Audio Support for the runtime.
 ## 1.1.0
-- Added require function
+- Added require function.
 - Added Chocolatey release and CI builds in Github actions.
-- First Appear of JoyDO.js, The first Library for Funser, with Display List Object engine.
+- First Appear of JoyDO.js, The first Library for Funser, with Display List Object and Audio engine support.
 ## 1.0.0
 - First release